% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_hdr.R
\name{hdr2df}
\alias{hdr2df}
\title{Convert `.hdr` file into data frame}
\usage{
hdr2df(path, species, pick_rank = NULL)
}
\arguments{
\item{path}{Path of the target hdr file}

\item{species}{Species in romaji}

\item{pick_rank}{`Rank` (A, B, C, etc.) to find location}
}
\description{
Convert `.hdr` file into data frame
}
\examples{
\dontrun{
  hdr2df("PATH_TO_HDR_FILE/sample.hdr", species = "maiwashi")
  hdr2df("PATH_TO_HDR_FILE/sample.hdr",
         species   = "maiwashi",
         pick_rank = "C")
}
}
