% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_hdr.R
\name{hdr2df_in_dir}
\alias{hdr2df_in_dir}
\title{Tidy all `.hdr`s into a single df in given directory}
\usage{
hdr2df_in_dir(dir, species, pick_rank = NULL)
}
\arguments{
\item{dir}{Target directory which contains `.hdr` files}

\item{species}{Species in romaji}

\item{pick_rank}{`Rank` (A, B, C, etc.) to find location}
}
\description{
Tidy all `.hdr`s into a single df in given directory
}
\examples{
\dontrun{
  hdr2df_in_dir("PATH_TO_HDR_FILE", species = "maiwashi")
}
}
