% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.otolith}
\alias{plot.otolith}
\title{Plot relationship between Age and OR}
\usage{
\method{plot}{otolith}(df_otolith)
}
\arguments{
\item{df_otolith}{Tidy otolith data made by \code{hdr2df_in_dir}}
}
\description{
Plot relationship between Age and OR
}
