% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{preview}
\alias{preview}
\title{Preview relationship between Age and OR}
\usage{
preview(df_otolith)
}
\arguments{
\item{df_otolith}{Tidy otolith data
made by \code{hdr2df} and \code{hdr2df_in_dir}}
}
\value{
Nothing. Side-effect: plots graphs
}
\description{
Preview relationship between Age and OR
}
